Namespace optimization
