+++ support/Db.php
@@ -14,6 +14,11 @@
 namespace support;
 
 use Illuminate\Database\Capsule\Manager;
+use Illuminate\Database\Capsule\Manager as Capsule;
+use Illuminate\Events\Dispatcher;
+use Illuminate\Container\Container;
+use Jenssegers\Mongodb\Connection;
+use Workerman\Timer;
 
 /**
  * Class Db
@@ -21,5 +26,92 @@ use Illuminate\Database\Capsule\Manager;
  */
 class Db extends Manager
 {
+    /**
+     * @return void
+     */
+    public static function setInstance()
+    {
+        $capsule = new Capsule;
+        $configs = config('database');
+        if (empty($configs)) {
+            return;
+        }
 
+        $capsule->getDatabaseManager()->extend('mongodb', function ($config, $name) {
+            $config['name'] = $name;
+            return new Connection($config);
+        });
+
+        $default_config = $configs['connections'][$configs['default']];
+        $capsule->addConnection($default_config);
+
+        foreach ($configs['connections'] as $name => $config) {
+            $capsule->addConnection($config, $name);
+        }
+
+        if (class_exists('\Illuminate\Events\Dispatcher')) {
+            $capsule->setEventDispatcher(new Dispatcher(new Container));
+        }
+
+        $capsule->setAsGlobal();
+
+        $capsule->bootEloquent();
+
+        // Heartbeat
+        $connections = config('database.connections');
+        if (!$connections) {
+            return;
+        }
+        Timer::add(55, function () use ($connections) {
+            foreach ($connections as $key => $item) {
+                if ($item['driver'] == 'mysql') {
+                    Db::connection($key)->select('select 1');
+                }
+            }
+        });
+    }
+
+    /**
+     * Get a connection instance from the global manager.
+     *
+     * @param  string|null  $connection
+     * @return \Illuminate\Database\Connection
+     */
+    public static function connection($connection = null)
+    {
+        if (!static::$instance) {
+            static::setInstance();
+        }
+        return static::$instance->getConnection($connection);
+    }
+
+    /**
+     * Get a fluent query builder instance.
+     *
+     * @param  \Closure|\Illuminate\Database\Query\Builder|string  $table
+     * @param  string|null  $as
+     * @param  string|null  $connection
+     * @return \Illuminate\Database\Query\Builder
+     */
+    public static function table($table, $as = null, $connection = null)
+    {
+        if (!static::$instance) {
+            static::setInstance();
+        }
+        return static::$instance->connection($connection)->table($table, $as);
+    }
+
+    /**
+     * Get a schema builder instance.
+     *
+     * @param  string|null  $connection
+     * @return \Illuminate\Database\Schema\Builder
+     */
+    public static function schema($connection = null)
+    {
+        if (!static::$instance) {
+            static::setInstance();
+        }
+        return static::$instance->connection($connection)->getSchemaBuilder();
+    }
 }
