+++ support/Log.php
@@ -11,9 +11,8 @@
  * @link      http://www.workerman.net/
  * @license   http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace support\bootstrap;
+namespace support;
 
-use Webman\Bootstrap;
 use Monolog\Logger;
 
 /**
@@ -30,40 +29,34 @@ use Monolog\Logger;
  * @method static void alert($message, array $context = [])
  * @method static void emergency($message, array $context = [])
  */
-class Log implements Bootstrap {
-
+class Log
+{
     /**
      * @var array
      */
     protected static $_instance = [];
 
     /**
-     * @param \Workerman\Worker $worker
-     * @return void
-     */
-    public static function start($worker)
-    {
-        $configs = config('log', []);
-        foreach ($configs as $channel => $config) {
-            $logger = static::$_instance[$channel] = new Logger($channel);
-            foreach ($config['handlers'] as $handler_config) {
-                $handler = new $handler_config['class'](... \array_values($handler_config['constructor']));
-                if (isset($handler_config['formatter'])) {
-                    $formatter = new $handler_config['formatter']['class'](... \array_values($handler_config['formatter']['constructor']));
-                    $handler->setFormatter($formatter);
-                }
-                $logger->pushHandler($handler);
-            }
-        }
-    }
-
-    /**
      * @param string $name
-     * @return Logger;
+     * @return Logger
      */
     public static function channel($name = 'default')
     {
-        return static::$_instance[$name] ?? null;
+        if (!static::$_instance) {
+            $configs = config('log', []);
+            foreach ($configs as $channel => $config) {
+                $logger = static::$_instance[$channel] = new Logger($channel);
+                foreach ($config['handlers'] as $handler_config) {
+                    $handler = new $handler_config['class'](... \array_values($handler_config['constructor']));
+                    if (isset($handler_config['formatter'])) {
+                        $formatter = new $handler_config['formatter']['class'](... \array_values($handler_config['formatter']['constructor']));
+                        $handler->setFormatter($formatter);
+                    }
+                    $logger->pushHandler($handler);
+                }
+            }
+        }
+        return static::$_instance[$name];
     }
