--- support/bootstrap/db/Laravel.php
@@ -1,77 +1,0 @@
-<?php
-/**
- * This file is part of webman.
- *
- * Licensed under The MIT License
- * For full copyright and license information, please see the MIT-LICENSE.txt
- * Redistributions of files must retain the above copyright notice.
- *
- * @author    walkor<walkor@workerman.net>
- * @copyright walkor<walkor@workerman.net>
- * @link      http://www.workerman.net/
- * @license   http://www.opensource.org/licenses/mit-license.php MIT License
- */
-namespace support\bootstrap\db;
-
-use Webman\Bootstrap;
-use Illuminate\Database\Capsule\Manager as Capsule;
-use Illuminate\Events\Dispatcher;
-use Illuminate\Container\Container;
-use Jenssegers\Mongodb\Connection;
-use Workerman\Timer;
-use Workerman\Worker;
-use support\Db;
-
-/**
- * Class Laravel
- * @package support\bootstrap\db
- */
-class Laravel implements Bootstrap
-{
-    /**
-     * @param Worker $worker
-     *
-     * @return void
-     */
-    public static function start($worker)
-    {
-        if (!class_exists('\Illuminate\Database\Capsule\Manager')) {
-            return;
-        }
-        $capsule = new Capsule;
-        $configs = config('database');
-        if (empty($configs)) {
-            return;
-        }
-
-        $capsule->getDatabaseManager()->extend('mongodb', function($config, $name) {
-            $config['name'] = $name;
-            return new Connection($config);
-        });
-
-        $default_config = $configs['connections'][$configs['default']];
-        $capsule->addConnection($default_config);
-
-        foreach ($configs['connections'] as $name => $config) {
-            $capsule->addConnection($config, $name);
-        }
-
-        if (class_exists('\Illuminate\Events\Dispatcher')) {
-            $capsule->setEventDispatcher(new Dispatcher(new Container));
-        }
-
-        $capsule->setAsGlobal();
-
-        $capsule->bootEloquent();
-
-        $connections = config('database.connections');
-        if (!$connections) {
-            return;
-        }
-        Timer::add(55, function () use ($connections){
-            foreach ($connections as $key => $item) {
-                Db::connection($key)->select('select 1 limit 1');
-            }
-        });
-    }
-}
